# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Latter House Life Planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /todos:
    get:
      operationId: listTodos
      tags: [trackers]
      summary: List todo items for the authenticated user
      parameters:
        - name: completed
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of todo items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoItem"
    post:
      operationId: createTodo
      tags: [trackers]
      summary: Create a todo item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTodoBody"
      responses:
        "201":
          description: Created todo item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoItem"
  /todos/{id}:
    put:
      operationId: updateTodo
      tags: [trackers]
      summary: Update a todo item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTodoBody"
      responses:
        "200":
          description: Updated todo item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoItem"
    delete:
      operationId: deleteTodo
      tags: [trackers]
      summary: Delete a todo item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /prayers:
    get:
      operationId: listPrayers
      tags: [trackers]
      summary: List prayer entries
      parameters:
        - name: month
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of prayer entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrayerEntry"
    post:
      operationId: createPrayer
      tags: [trackers]
      summary: Create a prayer entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrayerBody"
      responses:
        "201":
          description: Created prayer entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerEntry"
  /prayers/{id}:
    put:
      operationId: updatePrayer
      tags: [trackers]
      summary: Update a prayer entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePrayerBody"
      responses:
        "200":
          description: Updated prayer entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrayerEntry"
    delete:
      operationId: deletePrayer
      tags: [trackers]
      summary: Delete a prayer entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /habits:
    get:
      operationId: listHabits
      tags: [trackers]
      summary: List habits
      responses:
        "200":
          description: List of habits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Habit"
    post:
      operationId: createHabit
      tags: [trackers]
      summary: Create a habit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHabitBody"
      responses:
        "201":
          description: Created habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
  /habits/{id}:
    put:
      operationId: updateHabit
      tags: [trackers]
      summary: Update a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHabitBody"
      responses:
        "200":
          description: Updated habit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Habit"
    delete:
      operationId: deleteHabit
      tags: [trackers]
      summary: Delete a habit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /habits/logs:
    get:
      operationId: listHabitLogs
      tags: [trackers]
      summary: List habit logs for a month
      parameters:
        - name: month
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of habit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HabitLog"
  /habits/{id}/logs:
    post:
      operationId: logHabit
      tags: [trackers]
      summary: Log a habit completion for a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogHabitBody"
      responses:
        "201":
          description: Created habit log
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HabitLog"
  /habits/{id}/logs/{date}:
    delete:
      operationId: deleteHabitLog
      tags: [trackers]
      summary: Remove a habit log for a date
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: date
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
    TodoItem:
      type: object
      properties:
        id:
          type: integer
        text:
          type: string
        notes:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        completed:
          type: boolean
        completedAt:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, text, completed, createdAt, updatedAt]
    CreateTodoBody:
      type: object
      properties:
        text:
          type: string
        notes:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
      required: [text]
    UpdateTodoBody:
      type: object
      properties:
        text:
          type: string
        notes:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        completed:
          type: boolean
        completedAt:
          type: string
          nullable: true
        priority:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
    PrayerEntry:
      type: object
      properties:
        id:
          type: integer
        date:
          type: string
        title:
          type: string
          nullable: true
        body:
          type: string
        category:
          type: string
          nullable: true
        answered:
          type: boolean
        answeredDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, date, body, answered, createdAt, updatedAt]
    CreatePrayerBody:
      type: object
      properties:
        date:
          type: string
        title:
          type: string
          nullable: true
        body:
          type: string
        category:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [date, body]
    UpdatePrayerBody:
      type: object
      properties:
        title:
          type: string
          nullable: true
        body:
          type: string
        category:
          type: string
          nullable: true
        answered:
          type: boolean
        answeredDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
    Habit:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        frequency:
          type: string
        active:
          type: boolean
        createdAt:
          type: string
      required: [id, name, frequency, active, createdAt]
    CreateHabitBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        frequency:
          type: string
          nullable: true
      required: [name]
    UpdateHabitBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        emoji:
          type: string
          nullable: true
        color:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        frequency:
          type: string
        active:
          type: boolean
    HabitLog:
      type: object
      properties:
        id:
          type: integer
        habitId:
          type: integer
        date:
          type: string
        completed:
          type: boolean
        note:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, habitId, date, completed, createdAt]
    LogHabitBody:
      type: object
      properties:
        date:
          type: string
        note:
          type: string
          nullable: true
      required: [date]
...